refactor classes to appropriate packages
