+++ android/sdk/src/main/java/io/o2mc/sdk/business/batch/BatchDispatcher.java
@@ -1,4 +1,4 @@
-package io.o2mc.sdk.business;
+package io.o2mc.sdk.business.batch;
 
 import android.support.annotation.NonNull;
 import android.util.Log;
