+++ android/sdk/src/main/java/io/o2mc/sdk/business/batch/BatchGenerator.java
@@ -1,4 +1,4 @@
-package io.o2mc.sdk.business;
+package io.o2mc.sdk.business.batch;
 
 import android.util.Log;
