+++ android/sdk/src/main/java/io/o2mc/sdk/business/event/EventGenerator.java
@@ -1,4 +1,4 @@
-package io.o2mc.sdk.business;
+package io.o2mc.sdk.business.event;
 
 import android.util.Log;
